+++ web/src/admin.js
@@ -180,7 +180,6 @@ export function build_page() {
         can_invite_users_by_email: settings_data.user_can_invite_users_by_email(),
         realm_invite_required: realm.realm_invite_required,
         can_create_user_groups: settings_data.user_can_create_user_groups(),
-        policy_values: settings_config.common_policy_values,
         ...settings_org.get_organization_settings_options(),
         demote_inactive_streams_values: settings_config.demote_inactive_streams_values,
         web_mark_read_on_scroll_policy_values: