+++ web/src/settings_config.ts
@@ -227,29 +227,6 @@ export const email_address_visibility_values = {
     },
 };
 
-export const common_policy_values = {
-    by_admins_only: {
-        order: 1,
-        code: 2,
-        description: $t({defaultMessage: "Admins"}),
-    },
-    by_moderators_only: {
-        order: 2,
-        code: 4,
-        description: $t({defaultMessage: "Admins and moderators"}),
-    },
-    by_full_members: {
-        order: 3,
-        code: 3,
-        description: $t({defaultMessage: "Admins, moderators and full members"}),
-    },
-    by_members: {
-        order: 4,
-        code: 1,
-        description: $t({defaultMessage: "Admins, moderators and members"}),
-    },
-};
-
 export const email_invite_to_realm_policy_values = {
     nobody: {
         order: 1,